+++ powermatch/main.py
@@ -19,14 +19,20 @@
 #  <http://www.gnu.org/licenses/>.
 #
 # Note: Batch process is all rather messy.
-from copy import copy
-import os
-import sys
-from PyQt5 import QtCore, QtGui, QtWidgets
-import displayobject
-from displaytable import Table
+
 from credits import fileVersion
-from math import log10
+import configparser  # decode .ini file
+from copy import copy
+from core import (
+    AdjustmentsBase, Constraint, Facility, Optimisation, ProgressHandler, ProgressInfo,
+    PowerMatchBase, setTransitionBase,compute_lcoe, generate_summary
+    )
+from datetime import datetime
+from displaytable import Table
+import displayobject
+from editini import EdtDialog, SaveIni
+from floaters import ProgressBar, FloatStatus
+from getmodels import getModelFile, commonprefix
 import matplotlib
 if matplotlib.__version__ > '3.5.1':
     matplotlib.use('Qt5Agg')
@@ -37,17 +43,17 @@ import matplotlib.tri as mtri
 # This import registers the 3D projection, but is otherwise unused.
 from mpl_toolkits.mplot3d import Axes3D
 from matplotlib.ticker import LinearLocator, FormatStrFormatter
+import numpy as np
 import openpyxl as oxl
-from core import (Constraint, Facility, Optimisation, ProgressHandler, ProgressInfo,
-                             PowerMatchBase, setTransitionBase,compute_lcoe, generate_summary
-                             )
+import os
+from PyQt5 import QtCore, QtGui, QtWidgets
+import random
+import sys
 import shutil
 import subprocess
 from senutils import ClickableQLabel, getUser, ListWidget, ssCol, WorkBook
-from editini import EdtDialog, SaveIni
-from floaters import ProgressBar, FloatStatus
-from getmodels import getModelFile, commonprefix
-import configparser  # decode .ini file
+import time
+from typing import Optional
 from zoompan import ZoomPanX
 try:
     from opt_debug import optimiseDebug
@@ -55,53 +61,6 @@ except:
     pass
 
 from powermatch_lit import *
-
-def get_value(ws, row, col):
-    def get_range(text, alphabet=None, base=1):
-        if len(text) < 1:
-            return None
-        if alphabet is None:
-            alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
-        if alphabet[0] == ' ':
-            alphabet = alphabet[1:]
-        alphabet = alphabet.upper()
-        bits = ['', '']
-        b = 0
-        in_char = True
-        for char in text:
-            if char.isdigit():
-                if in_char:
-                    in_char = False
-                    b += 1
-            else:
-                if alphabet.find(char.upper()) < 0:
-                    continue
-                if not in_char:
-                    in_char = True
-                    b += 1
-            if b >= len(bits):
-                break
-            bits[b] += char.upper()
-        try:
-            bits[1] = int(bits[1]) - (1 - base)
-        except:
-            pass
-        row = 0
-        ndx = 1
-        for c in range(len(bits[0]) -1, -1, -1):
-            ndx1 = alphabet.index(bits[0][c]) + 1
-            row = row + ndx1 * ndx
-            ndx = ndx * len(alphabet)
-        bits[0] = row - (1 - base)
-        for c in bits:
-            if c == '':
-                return None
-    try:
-        while ws.cell(row=row, column=col).value[0] == '=':
-            row, col = get_range(ws.cell(row=row, column=col).value)
-    except:
-        return ''
-    return ws.cell(row=row, column=col).value
 
 class MyQDialog(QtWidgets.QDialog):
     ignoreEnter = True
@@ -122,36 +81,8 @@ class MyQDialog(QtWidgets.QDialog):
  #   def resizeEvent(self, event):
   #      return super(Window, self).resizeEvent(event)
 
-class Adjustments(MyQDialog):
-    def setAdjValueUnits(self, key, typ, capacity):
-        if key != 'Load':
-            mw = capacity
-            if typ == 'S':
-                unit = 'MWh'
-            else:
-                unit = 'MW'
-            dp = self._decpts
-            div = 0
-        else:
-            dimen = log10(capacity)
-            unit = 'MWh'
-            if dimen > 11:
-                unit = 'PWh'
-                div = 9
-            elif dimen > 8:
-                unit = 'TWh'
-                div = 6
-            elif dimen > 5:
-                unit = 'GWh'
-                div = 3
-            else:
-                div = 0
-            mw = capacity / pow(10, div)
-            dp = None
-        mwtxt = unit
-        mwcty = round(mw, dp)
-        return mw, mwtxt, mwcty, div
-
+class Adjustments(AdjustmentsBase, MyQDialog):
+    @staticmethod
     def niceSize(window, ctr): # works for Adjustments window (probably because less that 640*480)
         height = window.frameSize().height() / 1.07
         height = 65 + ctr * 32
@@ -166,35 +97,16 @@ class Adjustments(MyQDialog):
 
     def __init__(self, parent, data, adjustin, adjust_cap, prefix, show_multipliers=False, save_folder=None,
                  batch_file=None):
-        super(Adjustments, self).__init__()
-        self.ignoreEnter = False
-        self._adjust_typ = {} # facility type = G, S or L
-        self._adjust_cty = {} # (actual) adjust capacity
-        self.show_multipliers = show_multipliers
-        if self.show_multipliers:
-            self._adjust_mul = {} # (actual) adjust multiplier
-            self._adjust_rnd = {} # multiplier widget (rounded to 4 digits)
-        self._adjust_txt = {} # string with capacity units
-        self._save_folder = save_folder
-        self._batch_file = None
-        if batch_file is not None:
-            if os.path.isfile(batch_file):
-                self._batch_file = batch_file
-        self._ignore = False
-        self._results = None
+        AdjustmentsBase.__init__(self, parent, data, adjustin, adjust_cap, prefix, 
+                                show_multipliers=show_multipliers, 
+                                save_folder=save_folder, 
+                                batch_file=batch_file)
+        MyQDialog.__init__(self, parent)  # Initialize the QDialog properly
+        self.init_ui(data, adjustin, adjust_cap, prefix, save_folder)
+
+    def init_ui(self, data, adjustin, adjust_cap, prefix, save_folder):
         self.grid = QtWidgets.QGridLayout()
-        self._data = {}
         ctr = 0
-        self._decpts = 1
-        for key, typ, capacity in data:
-            if key == 'Load' or capacity is None:
-                continue
-            dimen = log10(capacity)
-            if dimen < 2.:
-                if dimen < 1.:
-                    self._decpts = 2
-                elif self._decpts != 2:
-                    self._decpts = 1
         if prefix is not None:
             self.grid.addWidget(QtWidgets.QLabel('Results Prefix:'), ctr, 0)
             self.pfx_fld = QtWidgets.QLineEdit()
@@ -292,75 +204,16 @@ class Adjustments(MyQDialog):
         self.scroll.setWidget(frame)
         self.layout = QtWidgets.QVBoxLayout(self)
         self.layout.addWidget(self.scroll)
-        self.niceSize(ctr)
+        # self.niceSize(ctr)
         self.setWindowTitle('SIREN - Powermatch - Adjust generators')
         self.setWindowIcon(QtGui.QIcon('sen_icon32.ico'))
         QtWidgets.QShortcut(QtGui.QKeySequence('q'), self, self.quitClicked)
         self.show()
-
-    def adjustMult(self):
-        key = self.sender().objectName()
-        if not self._ignore:
-            self._adjust_mul[key] = self._adjust_rnd[key].value()
-            self._adjust_cty[key].setValue(self._data[key][0] * self._adjust_rnd[key].value())
-        mw, mwtxt, mwstr, div = self.setAdjValueUnits(key, self._adjust_typ[key], self._data[key][0])
-        self._adjust_txt[key].setText(mwtxt)
-     #   if not self._ignore:
-      #      self._adjust_val[key].setText(mwstr)
-        self._ignore = False
-
-    def adjustCap(self):
-        if self._ignore:
-            return
-        key = self.sender().objectName()
-        if key != 'Load':
-            adj = self._adjust_cty[key].value() / self._data[key][0]
-         #   self._adjust_rnd[key].setValue(adj)
-        else:
-            dimen = log10(self._data[key][0])
-            if dimen > 11:
-                mul = 9
-            elif dimen > 8:
-                mul = 6
-            elif dimen > 5:
-                mul = 3
-            else:
-                mul = 0
-            adj = (self._adjust_cty[key].value() * pow(10, mul)) / self._data[key][0]
-        self._adjust_mul[key] = adj
-      #  self._adjust_cty[key] = self._data[key] * adj
-        self._ignore = True
-        self._adjust_rnd[key].setValue(round(adj, 4))
-        self._ignore = False
+        
 
     def quitClicked(self):
         self.ignoreEnter = False
         self.close()
-
-    def resetClicked(self, to):
-        if to is None:
-            to = 0.
-        else:
-            to = 1.
-        if self.show_multipliers:
-            for key in self._adjust_rnd.keys():
-                self._adjust_rnd[key].setValue(to)
-        else:
-            if to == 0:
-                for key in self._adjust_cty.keys():
-                    self._adjust_cty[key].setValue(0.)
-            else:
-                for key in self._adjust_cty.keys():
-                    self._adjust_cty[key].setValue(self._data[key][0])
-        self.pfx_fld.setText('')
-
-    def resetloadClicked(self, to):
-        if isinstance(to, bool):
-            to = 1.
-        if self.show_multipliers:
-            self._adjust_rnd['Load'].setValue(to)
-        else:
-            self._adjust_cty['Load'].setValue(self._data['Load'][0])
 
     def restoreClicked(self):
         ini_file = QtWidgets.QFileDialog.getOpenFileName(self, 'Open Adjustments file',
@@ -375,28 +228,6 @@ class Adjustments(MyQDialog):
             except:
                 prefix = ''
             self.getIt(config, prefix)
-
-    def getIt(self, config, prefix=''):
-        try:
-            adjustto = config.get('Powermatch', 'adjusted_capacities')
-        except:
-            return
-        self.resetClicked(to=None)
-        bits = adjustto.split(',')
-        for bit in bits:
-            bi = bit.split('=')
-            key = bi[0]
-            try:
-                mw, mwtxt, mwcty, div = self.setAdjValueUnits(key, self._adjust_typ[key],
-                                        float(bi[1]))
-                self._adjust_cty[key].setValue(mwcty)
-                if self.show_multipliers:
-                    self._adjust_mul[key] = float(bi[1]) / (self._data[key][0] * pow(10, self._data[key][1]))
-                    self._adjust_rnd[key].setValue(round(self._adjust_mul[key], 4))
-            except:
-                pass
-        self._ignore = False
-        self.pfx_fld.setText(prefix)
 
     def listClicked(self):
         if os.path.exists(self._save_folder):
@@ -469,19 +300,6 @@ class Adjustments(MyQDialog):
                 if inifile[-4:] != '.ini':
                     inifile = inifile + '.ini'
                 SaveIni(updates, ini_file=inifile)
-
-    def showClicked(self):
-        self.ignoreEnter = False
-        self._results = {}
-        for key in list(self._adjust_cty.keys()):
-            self._results[key] = self._adjust_cty[key].value() * pow(10, self._data[key][1])
-        self.close()
-
-    def getValues(self):
-        return self._results
-
-    def getPrefix(self):
-        return self.pfx_fld.text()
 
     def addtoBatch(self):
         check_list = list(self._adjust_cty.keys())[1:]
@@ -603,7 +421,8 @@ class setTransition(setTransitionBase, MyQDialog):
         window.resize(size)
 
     def __init__(self, parent, label, generators, sheet, year):
-        super(setTransition, self).__init__()
+        setTransitionBase.__init__(self, parent, label, generators, sheet, year)
+        MyQDialog.__init__(self, parent)  # Initialize the QDialog properly
         self._results = None
         i = generators.rfind('/')
         generator_file = generators[i + 1:]
@@ -651,15 +470,13 @@ class setTransition(setTransitionBase, MyQDialog):
         self._results = self.sheet.text()
         self.close()
 
-    def getValues(self):
-        return self._results
-
 class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
     log = QtCore.pyqtSignal()
     progress = QtCore.pyqtSignal()
 
     def __init__(self, TableClass=Table, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+        QtWidgets.QWidget.__init__(self)
+        PowerMatchBase.__init__(self)
         self.displaytable = TableClass  # Store the TableClass as an instance variable
         self.help = help
         config = configparser.RawConfigParser()
@@ -721,11 +538,11 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
                     self.sheets[i].setCurrentIndex(ndx)
                     self.ws = ts.sheet_by_index(ndx)
                     if i == C:
-                        self.constraints = self.getConstraints()
+                        self.constraints = self.getConstraints(self.ws)
                     elif i == G:
-                        self.generators = self.getGenerators()
+                        self.generators = self.getGenerators(self.ws)
                     elif i == O:
-                        self.optimisation = self.getOptimisation()
+                        self.optimisation = self.getOptimisation(self.ws)
                     ts.close()
                     del ts
                 except Exception as error:
@@ -849,15 +666,15 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
         btch = QtWidgets.QPushButton('Batch', self)
         self.grid.addWidget(btch, r, 3)
         # btch.clicked.connect(self.pmClicked)
-        btch.clicked.connect(lambda checked, handler=self.progress_handler: self.pmClicked('Batch', handler))
+        btch.clicked.connect(lambda checked, handler=self.progress_handler, adjuster=self.create_adjustments: self.pmClicked('Batch', handler, adjuster))
         trns = QtWidgets.QPushButton('Transition', self)
         self.grid.addWidget(trns, r, 4)
         # trns.clicked.connect(self.pmClicked)
-        trns.clicked.connect(lambda checked, handler=self.progress_handler: self.pmClicked('Transition', handler))
+        trns.clicked.connect(lambda checked, handler=self.progress_handler, set_transition=self.create_setTransitions: self.pmClicked('Transition', handler, None, set_transition))
         opt = QtWidgets.QPushButton('Optimise', self)
         self.grid.addWidget(opt, r, 5)
         # opt.clicked.connect(self.pmClicked)
-        opt.clicked.connect(lambda checked, handler=self.progress_handler: self.pmClicked('Optimise', handler))
+        opt.clicked.connect(lambda checked, handler=self.progress_handler, optimiser=self.optClicked: self.pmClicked('Optimise', handler, optimiser))
         help = QtWidgets.QPushButton('Help', self)
      #   help.setMaximumWidth(wdth)
       #  quit.setMaximumWidth(wdth)
@@ -940,6 +757,1315 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
         else:
             self.center()
             self.resize(int(self.sizeHint().width() * 1.2), int(self.sizeHint().height() * 1.2))
+
+    def create_adjustments(self, pmss_data, load_col):
+        if self.adjust_gen:
+            generated = sum(pmss_data[load_col])
+            datain = [['Load', 'L', generated]]
+            adjustto = {'Load': generated}
+            adjust = Adjustments(self, datain, adjustto, self.adjust_cap, None,
+                                        show_multipliers=self.show_multipliers)
+            adjust.exec_()
+            return adjust
+        return None
+    
+    def create_setTransitions(self, label, generators, sheet, year):
+        files = setTransition(self, label, generators,
+                                      sheet, year)
+        files.exec_()
+        return files
+
+    def optClicked(self, in_year, sender_text, in_option, in_pmss_details, in_pmss_data, in_re_order,
+                   in_dispatch_order, pm_data_file, data_file, opt_progress_handler: Optional[ProgressHandler] = None, ):
+
+        def create_starting_population(individuals, chromosome_length):
+            # Set up an initial array of all zeros
+            population = np.zeros((individuals, chromosome_length))
+            # Loop through each row (individual)
+            for i in range(individuals):
+                # Choose a random number of ones to create but at least one 1
+                ones = random.randint(1, chromosome_length)
+                # Change the required number of zeros to ones
+                population[i, 0:ones] = 1
+                # Sfuffle row
+                np.random.shuffle(population[i])
+            return population
+
+        def select_individual_by_tournament(population, *argv):
+            # Get population size
+            population_size = len(population)
+
+            # Pick individuals for tournament
+            fighter = [0, 0]
+            fighter_fitness = [0, 0]
+            fighter[0] = random.randint(0, population_size - 1)
+            fighter[1] = random.randint(0, population_size - 1)
+
+            # Get fitness score for each
+            if len(argv) == 1:
+                fighter_fitness[0] = argv[0][fighter[0]]
+                fighter_fitness[1] = argv[0][fighter[1]]
+            else:
+                for arg in argv:
+                    min1 = min(arg)
+                    max1 = max(arg)
+                    for f in range(len(fighter)):
+                        try:
+                            fighter_fitness[f] += (arg[f] - min1) / (max1 - min1)
+                        except:
+                            pass
+            # Identify individual with lowest score
+            # Fighter 1 will win if score are equal
+            if fighter_fitness[0] <= fighter_fitness[1]:
+                winner = fighter[0]
+            else:
+                winner = fighter[1]
+
+            # Return the chromsome of the winner
+            return population[winner, :]
+
+        def breed_by_crossover(parent_1, parent_2, points=2):
+            # Get length of chromosome
+            chromosome_length = len(parent_1)
+
+            # Pick crossover point, avoiding ends of chromsome
+            if points == 1:
+                crossover_point = random.randint(1,chromosome_length-1)
+            # Create children. np.hstack joins two arrays
+                child_1 = np.hstack((parent_1[0:crossover_point],
+                                     parent_2[crossover_point:]))
+                child_2 = np.hstack((parent_2[0:crossover_point],
+                                     parent_1[crossover_point:]))
+            else: # only do 2 at this    stage
+                crossover_point_1 = random.randint(1, chromosome_length - 2)
+                crossover_point_2 = random.randint(crossover_point_1 + 1, chromosome_length - 1)
+                child_1 = np.hstack((parent_1[0:crossover_point_1],
+                                     parent_2[crossover_point_1:crossover_point_2],
+                                     parent_1[crossover_point_2:]))
+                child_2 = np.hstack((parent_2[0:crossover_point_1],
+                                     parent_1[crossover_point_1:crossover_point_2],
+                                     parent_2[crossover_point_2:]))
+            # Return children
+            return child_1, child_2
+
+        def randomly_mutate_population(population, mutation_probability):
+            # Apply random mutation
+            random_mutation_array = np.random.random(size=(population.shape))
+            random_mutation_boolean = random_mutation_array <= mutation_probability
+        #    random_mutation_boolean[0][:] = False # keep the best multi and lcoe
+       #     random_mutation_boolean[1][:] = False
+            population[random_mutation_boolean] = np.logical_not(population[random_mutation_boolean])
+            # Return mutation population
+            return population
+
+        def calculate_fitness(population):
+            lcoe_fitness_scores = [] # scores = LCOE values
+            multi_fitness_scores = [] # scores = multi-variable weight
+            multi_values = [] # values for each of the six variables
+            if len(population) == 1:
+                option = O1
+            else:
+                option = O
+            if self.debug:
+                self.popn += 1
+                self.chrom = 0
+            for chromosome in population:
+                # now get random amount of generation per technology (both RE and non-RE)
+                for fac, value in opt_order.items():
+                    capacity = value[2]
+                    for c in range(value[0], value[1]):
+                        if chromosome[c]:
+                            capacity = capacity + capacities[c]
+                    try:
+                        pmss_details[fac].multiplier = capacity / pmss_details[fac].capacity
+                    except:
+                        print('PME2:', gen, capacity, pmss_details[fac].capacity)
+                multi_value, op_data, extra = self.doDispatch(year, sender_text, option, pmss_details, pmss_data, re_order,
+                                              dispatch_order, pm_data_file, data_file, progress_handler=opt_progress_handler)
+                if multi_value['load_pct'] < self.targets['load_pct'][3]:
+                    if multi_value['load_pct'] == 0:
+                        print('PME3:', multi_value['lcoe'], self.targets['load_pct'][3], multi_value['load_pct'])
+                        lcoe_fitness_scores.append(1)
+                    else:
+                        try:
+                            lcoe_fitness_scores.append(pow(multi_value['lcoe'],
+                                self.targets['load_pct'][3] / multi_value['load_pct']))
+                        except OverflowError as err:
+                            self.setStatus(f"Overflow error: {err}; POW({multi_value['lcoe']:,}, " \
+                                         + f"{self.targets['load_pct'][3] / multi_value['load_pct']:,}) " \
+                                         + f"({self.targets['load_pct'][3]:,} / {multi_value['load_pct']:,} )")
+                        except:
+                            pass
+                else:
+                    lcoe_fitness_scores.append(multi_value['lcoe'])
+                multi_values.append(multi_value)
+                multi_fitness_scores.append(calc_weight(multi_value))
+                if self.debug:
+                    self.chrom += 1
+                    line = str(self.popn) + ',' + str(self.chrom) + ','
+                    for fac, value in opt_order.items():
+                        try:
+                            line += str(pmss_details[fac].capacity * pmss_details[fac].multiplier) + ','
+                        except:
+                            line += ','
+                    for key in self.targets.keys():
+                        try:
+                            line += '{:.3f},'.format(multi_value[key])
+                        except:
+                            line += multi_value[key] + ','
+                    line += '{:.5f},'.format(multi_fitness_scores[-1])
+                    self.db_file.write(line + '\n')
+            # alternative approach to calculating fitness
+            multi_fitness_scores1 = []
+            maxs = {}
+            mins = {}
+            tgts = {}
+            for key in multi_value.keys():
+                if key[-4:] == '_pct':
+                    tgts[key] = abs(self.targets[key][2] - self.targets[key][3])
+                else:
+                    maxs[key] = 0
+                    mins[key] = -1
+                    for popn in multi_values:
+                        try:
+                            tgt = abs(self.targets[key][2] - popn[key])
+                        except:
+                            continue
+                        if tgt > maxs[key]:
+                            maxs[key] = tgt
+                        if mins[key] < 0 or tgt < mins[key]:
+                            mins[key] = tgt
+            for popn in multi_values:
+                weight = 0
+                for key, value in multi_value.items():
+                    if self.targets[key][1] <= 0:
+                        continue
+                    try:
+                        tgt = abs(self.targets[key][2] - popn[key])
+                    except:
+                        continue
+                    if key[-4:] == '_pct':
+                        if tgts[key] != 0:
+                            if tgt > tgts[key]:
+                                weight += 1 * self.targets[key][1]
+                            else:
+                                try:
+                                    weight += 1 - ((tgt / tgts[key]) * self.targets[key][1])
+                                except:
+                                    pass
+                    else:
+                        try:
+                            weight += 1 - (((maxs[key] - tgt) / (maxs[key] - mins[key])) \
+                                      * self.targets[key][1])
+                        except:
+                            pass
+                multi_fitness_scores1.append(weight)
+            if len(population) == 1: # return the table for best chromosome
+                return op_data, multi_values
+            else:
+                return lcoe_fitness_scores, multi_fitness_scores, multi_values
+
+        def optQuitClicked(event):
+            self.optExit = True
+            optDialog.close()
+
+        def chooseClicked(event):
+            self.opt_choice = sender_text
+            chooseDialog.close()
+
+        def calc_weight(multi_value, calc=0):
+            weight = [0., 0.]
+            if calc == 0:
+                for key, value in self.targets.items():
+                    if multi_value[key] == '':
+                        continue
+                    if value[1] <= 0:
+                        continue
+                    if value[2] == value[3]: # wants specific target
+                        if multi_value[key] == value[2]:
+                            w = 0
+                        else:
+                            w = 1
+                    elif value[2] > value[3]: # wants higher target
+                        if multi_value[key] > value[2]: # high no weight
+                            w = 0.
+                        elif multi_value[key] < value[3]: # low maximum weight
+                            w = 2.
+                        else:
+                            w = 1 - (multi_value[key] - value[3]) / (value[2] - value[3])
+                    else: # lower target
+                        if multi_value[key] == -1 or multi_value[key] > value[3]: # high maximum weight
+                            w = 2.
+                        elif multi_value[key] < value[2]: # low no weight
+                            w = 0.
+                        else:
+                            w = multi_value[key] / (value[3] - value[2])
+                    weight[0] += w * value[1]
+            elif calc == 1:
+                for key, value in self.targets.items():
+                    if multi_value[key] == '':
+                        continue
+                    if value[1] <= 0:
+                        continue
+                    if multi_value[key] < 0:
+                        w = 1
+                    elif value[2] == value[3]: # wants specific target
+                        if multi_value[key] == value[2]:
+                            w = 0
+                        else:
+                            w = 1
+                    else: # target range
+                        w = min(abs(value[2] - multi_value[key]) / abs(value[2] - value[3]), 1)
+                    weight[1] += w * value[1]
+            return weight[calc]
+
+        def show_multitable(best_score_progress, multi_best, multi_order, title):
+            def pwr_chr(amt):
+                pwr_chrs = ' KMBTPEZY'
+                pchr = ''
+                divisor = 1.
+                for pwr in range(len(pwr_chrs) - 1, -1, -1):
+                    if amt > pow(10, pwr * 3):
+                        pchr = pwr_chrs[pwr]
+                        divisor = 1. * pow(10, pwr * 3)
+                        break
+                return amt / divisor, pchr
+
+            def opt_fmat(amt, fmat):
+                tail = ' '
+                p = fmat.find('pwr_chr')
+                if p > 0:
+                    amt, tail = pwr_chr(amt)
+                    fmat = fmat.replace('pwr_chr', '')
+              #  d = fmat.find('d')
+             #   if d > 0:
+              #      amt = amt * 100.
+                fmat = fmat.replace('.1f', '.2f')
+                i = fmat.find('%')
+                fmt = fmat[:i] + '{:> 8,' + fmat[i:].replace('%', '').replace('d', '.2%') + '}' + tail
+                return fmt.format(amt)
+
+            best_table = []
+            best_fmate = []
+            for b in range(len(multi_best)):
+                bl = list(multi_best[b].values())
+                bl.insert(0, best_score_progress[b])
+                bl.insert(0, b + 1)
+                best_table.append(bl)
+                best_fmate.append([b + 1])
+                best_fmate[-1].insert(1, best_score_progress[b])
+                for f in range(2, len(best_table[-1])):
+                    if target_keys[f - 2] in ['load_pct', 'surplus_pct', 're_pct']:
+                        best_fmate[-1].append(opt_fmat(bl[f], '%d%%'))
+                    else:
+                        best_fmate[-1].append(opt_fmat(bl[f], target_fmats[f - 2]))
+            fields = target_names[:]
+            fields.insert(0, 'weight')
+            fields.insert(0, 'iteration')
+            dialog = self.displaytable.Table(best_fmate, fields=fields, txt_align='R', decpts=[0, 4],
+                     title=title, sortby='weight')
+            dialog.exec_()
+            b = int(dialog.getItem(0)) - 1
+            del dialog
+            pick = [b]
+            for fld in multi_order[:3]:
+                i = target_keys.index(fld)
+                pick.append(best_table[b][i + 2])
+            return [pick]
+
+#       optClicked mainline starts here
+        year = in_year
+        option = in_option
+        pmss_details = dict(in_pmss_details)
+        pmss_data = in_pmss_data[:]
+        re_order = in_re_order[:]
+        dispatch_order = in_dispatch_order[:]
+        if self.optimise_debug:
+            self.debug = True
+        else:
+            self.debug = False
+        missing = []
+        for fac in re_order:
+            if fac == 'Load':
+                continue
+            if fac not in self.optimisation.keys():
+                missing.append(fac)
+        for gen in dispatch_order:
+            if gen not in self.optimisation.keys():
+                missing.append(gen)
+        if len(missing) > 0:
+            bad = False
+            if self.optimise_default is not None:
+                defaults = self.optimise_default.split(',')
+                if len(defaults) < 1 or len(defaults) > 3:
+                    bad = True
+                else:
+                    try:
+                        for miss in missing:
+                            if len(defaults) == 2:
+                                minn = 0
+                            else:
+                                if defaults[0][-1] == 'd':
+                                    minn = pmss_details[miss].capacity * float(defaults[0][:-1])
+                                elif defaults[0][-1] == 'c':
+                                    minn = pmss_details[miss].capacity * pmss_details[miss].multiplier * float(defaults[0][:-1])
+                                else:
+                                    minn = float(defaults[0])
+                            if defaults[-2][-1] == 'd':
+                                maxx = pmss_details[miss].capacity * float(defaults[-2][:-1])
+                            elif defaults[-2][-1] == 'c':
+                                maxx = pmss_details[miss].capacity * pmss_details[miss].multiplier * float(defaults[-2][:-1])
+                            else:
+                                maxx = float(defaults[-2])
+                            if len(defaults) == 3:
+                                if defaults[-1][-1].lower() == 'd':
+                                    step = capacity * float(defaults[-1][:-1])
+                                elif defaults[-1][-1].lower() == 'c':
+                                    step = capacity * multiplier * float(defaults[-1][:-1])
+                                else:
+                                    step = float(defaults[-1])
+                            else:
+                                step = (maxx - minn) / float(defaults[-1])
+                            self.optimisation[miss] =  Optimisation(miss, 'None', None)
+                            self.optimisation[miss].approach = 'Range'
+                            self.optimisation[miss].capacity_min = minn
+                            self.optimisation[miss].capacity_max = maxx
+                            self.optimisation[miss].capacity_step = step
+                    except:
+                        bad = True
+                check = ''
+                for miss in missing:
+                    check += miss + ', '
+                check = check[:-2]
+                if bad:
+                    self.setStatus('Key Error: Missing Optimisation entries for: ' + check)
+                    return
+                self.setStatus('Missing Optimisation entries added for: ' + check)
+        self.optExit = False
+        self.setStatus('Optimise processing started')
+        err_msg = ''
+        
+        optDialog = QtWidgets.QDialog()
+        grid = QtWidgets.QGridLayout()
+        grid.addWidget(QtWidgets.QLabel('Adjust load'), 0, 0)
+        self.optLoad = QtWidgets.QDoubleSpinBox()
+        self.optLoad.setRange(-1, self.adjust_cap)
+        self.optLoad.setDecimals(4)
+        self.optLoad.setSingleStep(.1)
+        rw = 0
+        grid.addWidget(self.optLoad, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Multiplier for input Load'), rw, 2, 1, 3)
+        self.optLoad.setValue(pmss_details['Load'].multiplier)
+        rw += 1
+        grid.addWidget(QtWidgets.QLabel('Population size'), rw, 0)
+        optPopn = QtWidgets.QSpinBox()
+        optPopn.setRange(10, 500)
+        optPopn.setSingleStep(10)
+        optPopn.setValue(self.optimise_population)
+        optPopn.valueChanged.connect(self.changes)
+        grid.addWidget(optPopn, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Size of population'), rw, 2, 1, 3)
+        rw += 1
+        grid.addWidget(QtWidgets.QLabel('No. of iterations'), rw, 0, 1, 3)
+        optGenn = QtWidgets.QSpinBox()
+        optGenn.setRange(10, 500)
+        optGenn.setSingleStep(10)
+        optGenn.setValue(self.optimise_generations)
+        optGenn.valueChanged.connect(self.changes)
+        grid.addWidget(optGenn, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Number of iterations (generations)'), rw, 2, 1, 3)
+        rw += 1
+        grid.addWidget(QtWidgets.QLabel('Mutation probability'), rw, 0)
+        optMutn = QtWidgets.QDoubleSpinBox()
+        optMutn.setRange(0, 1)
+        optMutn.setDecimals(4)
+        optMutn.setSingleStep(0.001)
+        optMutn.setValue(self.optimise_mutation)
+        optMutn.valueChanged.connect(self.changes)
+        grid.addWidget(optMutn, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Add in mutation'), rw, 2, 1, 3)
+        rw += 1
+        grid.addWidget(QtWidgets.QLabel('Exit if stable'), rw, 0)
+        optStop = QtWidgets.QSpinBox()
+        optStop.setRange(0, 50)
+        optStop.setSingleStep(10)
+        optStop.setValue(self.optimise_stop)
+        optStop.valueChanged.connect(self.changes)
+        grid.addWidget(optStop, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Exit if LCOE/weight remains the same after this many iterations'),
+                       rw, 2, 1, 3)
+        rw += 1
+        grid.addWidget(QtWidgets.QLabel('Optimisation choice'), rw, 0)
+        optCombo = QtWidgets.QComboBox()
+        choices = ['LCOE', 'Multi', 'Both']
+        for choice in choices:
+            optCombo.addItem(choice)
+            if choice == self.optimise_choice:
+                optCombo.setCurrentIndex(optCombo.count() - 1)
+        grid.addWidget(optCombo, rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Choose type of optimisation'),
+                       rw, 2, 1, 3)
+        rw += 1
+        # for each variable name
+        grid.addWidget(QtWidgets.QLabel('Variable'), rw, 0)
+        grid.addWidget(QtWidgets.QLabel('Weight'), rw, 1)
+        grid.addWidget(QtWidgets.QLabel('Better'), rw, 2)
+        grid.addWidget(QtWidgets.QLabel('Worse'), rw, 3)
+        rw += 1
+        ndx = grid.count()
+        for key in self.targets.keys():
+            self.targets[key][4] = ndx
+            ndx += 4
+        for key, value in self.targets.items():
+            if value[2] == value[3]:
+                ud = '(=)'
+            elif value[2] < 0:
+                ud = '(<html>&uarr;</html>)'
+            elif value[3] < 0 or value[3] > value[2]:
+                ud = '(<html>&darr;</html>)'
+            else:
+                ud = '(<html>&uarr;</html>)'
+            grid.addWidget(QtWidgets.QLabel(value[0] + ': ' + ud), rw, 0)
+            weight = QtWidgets.QDoubleSpinBox()
+            weight.setRange(0, 1)
+            weight.setDecimals(2)
+            weight.setSingleStep(0.05)
+            weight.setValue(value[1])
+            grid.addWidget(weight, rw, 1)
+            if key[-4:] == '_pct':
+                minim = QtWidgets.QDoubleSpinBox()
+                minim.setRange(-.1, 1.)
+                minim.setDecimals(2)
+                minim.setSingleStep(0.1)
+                minim.setValue(value[2])
+                grid.addWidget(minim, rw, 2)
+                maxim = QtWidgets.QDoubleSpinBox()
+                maxim.setRange(-.1, 1.)
+                maxim.setDecimals(2)
+                maxim.setSingleStep(0.1)
+                maxim.setValue(value[3])
+                grid.addWidget(maxim, rw, 3)
+            else:
+                minim = QtWidgets.QLineEdit()
+                minim.setValidator(QtGui.QDoubleValidator())
+                minim.validator().setDecimals(2)
+                minim.setText(str(value[2]))
+                grid.addWidget(minim, rw, 2)
+                maxim = QtWidgets.QLineEdit()
+                maxim.setValidator(QtGui.QDoubleValidator())
+                maxim.validator().setDecimals(2)
+                maxim.setText(str(value[3]))
+                grid.addWidget(maxim, rw, 3)
+            rw += 1
+        quit = QtWidgets.QPushButton('Quit', self)
+        grid.addWidget(quit, rw, 0)
+        quit.clicked.connect(self._optQuitClicked)
+        show = QtWidgets.QPushButton('Proceed', self)
+        grid.addWidget(show, rw, 1)
+        show.clicked.connect(optDialog.close)
+        optDialog.setLayout(grid)
+        optDialog.setWindowTitle('Choose Optimisation Parameters')
+        optDialog.setWindowIcon(QtGui.QIcon('sen_icon32.ico'))
+        optDialog.exec_()
+        if self.optExit: # a fudge to exit
+            self.setStatus('Execution aborted.')
+            return
+        # check we have optimisation entries for generators and storage
+        # update any changes to targets
+        self.optimise_choice = optCombo.currentText()
+        for key in self.targets.keys():
+            weight = grid.itemAt(self.targets[key][4] + 1).widget()
+            self.targets[key][1] = weight.value()
+            minim = grid.itemAt(self.targets[key][4] + 2).widget()
+            try:
+                self.targets[key][2] = minim.value()
+            except:
+                self.targets[key][2] = float(minim.text())
+            maxim = grid.itemAt(self.targets[key][4] + 3).widget()
+            try:
+                self.targets[key][3] = maxim.value()
+            except:
+                self.targets[key][3] = float(maxim.text())
+        # might want to save load value if changed
+        pmss_details['Load'].multiplier = self.optLoad.value()
+    
+        updates = {}
+        lines = []
+        lines.append('optimise_choice=' + self.optimise_choice)
+        lines.append('optimise_generations=' + str(self.optimise_generations))
+        lines.append('optimise_mutation=' + str(self.optimise_mutation))
+        lines.append('optimise_population=' + str(self.optimise_population))
+        lines.append('optimise_stop=' + str(self.optimise_stop))
+        if self.optimise_choice == 'LCOE':
+            do_lcoe = True
+            do_multi = False
+        elif self.optimise_choice == 'Multi':
+            do_lcoe = False
+            do_multi = True
+        else:
+            do_lcoe = True
+            do_multi = True
+        multi_order = []
+        for key, value in self.targets.items():
+            line = 'optimise_{}={:.2f},{:.2f},{:.2f}'.format(key, value[1], value[2], value[3])
+            lines.append(line)
+            multi_order.append('{:.2f}{}'.format(value[1], key))
+        updates['Powermatch'] = lines
+        SaveIni(updates)
+        multi_order.sort(reverse=True)
+    #    multi_order = multi_order[:3] # get top three weighted variables - but I want them all
+        multi_order = [o[4:] for o in multi_order]
+        self.adjust_gen = True
+        orig_load = []
+        load_col = -1
+        orig_tech = {}
+        orig_capacity = {}
+        opt_order = {} # rely on it being processed in added order
+        # each entry = [first entry in chrom, last entry, minimum capacity]
+        # first get original renewables generation from data sheet
+        for fac in re_order:
+            if fac == 'Load':
+                continue
+            opt_order[fac] = [0, 0, 0]
+        # now add scheduled generation
+        for gen in dispatch_order:
+            opt_order[gen] = [0, 0, 0]
+        capacities = []
+        for gen in opt_order.keys():
+            opt_order[gen][0] = len(capacities) # first entry
+            try:
+                if self.optimisation[gen].approach == 'Discrete':
+                    capacities.extend(self.optimisation[gen].capacities)
+                    opt_order[gen][1] = len(capacities) # last entry
+                elif self.optimisation[gen].approach == 'Range':
+                    if self.optimisation[gen].capacity_max == self.optimisation[gen].capacity_min:
+                        capacities.extend([0])
+                        opt_order[gen][1] = len(capacities)
+                        opt_order[gen][2] = self.optimisation[gen].capacity_min
+                        continue
+                    ctr = int((self.optimisation[gen].capacity_max - self.optimisation[gen].capacity_min) / \
+                              self.optimisation[gen].capacity_step)
+                    if ctr < 1:
+                        self.setStatus("Error with Optimisation table entry for '" + gen + "'")
+                        return
+                    capacities.extend([self.optimisation[gen].capacity_step] * ctr)
+                    tot = self.optimisation[gen].capacity_step * ctr + self.optimisation[gen].capacity_min
+                    if tot < self.optimisation[gen].capacity_max:
+                        capacities.append(self.optimisation[gen].capacity_max - tot)
+                    opt_order[gen][1] = len(capacities)
+                    opt_order[gen][2] = self.optimisation[gen].capacity_min
+                else:
+                    opt_order[gen][1] = len(capacities)
+            except KeyError as err:
+                self.setStatus('Key Error: No Optimisation entry for ' + str(err))
+                opt_order[gen] = [len(capacities), len(capacities) + 5, 0]
+                capacities.extend([pmss_details[gen].capacity / 5.] * 5)
+            except ZeroDivisionError as err:
+                self.setStatus('Zero capacity: ' + gen + ' ignored')
+            except:
+                err = str(sys.exc_info()[0]) + ',' + str(sys.exc_info()[1]) + ',' + gen + ',' \
+                      + str(opt_order[gen])
+                self.setStatus('Error: ' + str(err))
+                return
+        # chromosome = [1] * int(len(capacities) / 2) + [0] * (len(capacities) - int(len(capacities) / 2))
+        # we have the original data - from here down we can do our multiple optimisations
+        # Set general parameters
+        self.setStatus('Optimisation choice is ' + self.optimise_choice)
+        chromosome_length = len(capacities)
+        self.setStatus(f'Chromosome length: {chromosome_length}; {pow(2, chromosome_length):,} permutations')
+        self.optimise_population = optPopn.value()
+        population_size = self.optimise_population
+        self.optimise_generations = optGenn.value()
+        maximum_generation = self.optimise_generations
+        self.optimise_mutation = optMutn.value()
+        self.optimise_stop = optStop.value()
+        lcoe_scores = []
+        multi_scores = []
+        multi_values = []
+     #   if do_lcoe:
+      #      lcoe_target = 0. # aim for this LCOE
+        if do_multi:
+            multi_best = [] # list of six variables for best weight
+            multi_best_popn = [] # list of chromosomes for best weight
+        self.show_ProgressBar(maximum=optGenn.value(), msg='Process iterations', title='SIREN - Powermatch Progress')
+        self.opt_progressbar.setVisible(True)
+        start_time = time.time()
+        # Create starting population
+        self.opt_progressbar.barProgress(1, 'Processing iteration 1')
+        QtCore.QCoreApplication.processEvents()
+        population = create_starting_population(population_size, chromosome_length)
+        # calculate best score(s) in starting population
+        # if do_lcoe best_score = lowest non-zero lcoe
+        # if do_multi best_multi = lowest weight and if not do_lcoe best_score also = best_weight
+        if self.debug:
+            filename = self.scenarios + 'opt_debug_' + \
+                datetime.now().strftime('_%Y-%M-%d_%H%M') + '.csv'
+            self.db_file = open(filename, 'w')
+            line0 = 'Popn,Chrom,'
+            line1 = 'Weights,,'
+            line2 = 'Targets,,'
+            line3 = 'Range,' + str(population_size) + ','
+            for gen, value in opt_order.items():
+                 line0 += gen + ','
+                 line1 += ','
+                 line2 += ','
+                 line3 += ','
+            for key in self.targets.keys():
+                 line0 += key + ','
+                 line1 += str(self.targets[key][1]) + ','
+                 line2 += str(self.targets[key][2]) + ','
+                 if key[-4:] == '_pct':
+                     line3 += str(abs(self.targets[key][2] - self.targets[key][3])) + ','
+                 else:
+                     line3 += ','
+            line0 += 'Score'
+            self.db_file.write(line0 + '\n' + line1 + '\n' + line2 + '\n' + line3 + '\n')
+            self.popn = 0
+            self.chrom = 0
+        lcoe_scores, multi_scores, multi_values = calculate_fitness(population)
+        if do_lcoe:
+            try:
+                best_score = np.min(lcoe_scores)
+            except:
+                print('PME4:', lcoe_scores)
+            best_ndx = lcoe_scores.index(best_score)
+            lowest_chrom = population[best_ndx]
+            self.setStatus('Starting LCOE: $%.2f' % best_score)
+        if do_multi:
+            if self.more_details: # display starting population ?
+                pick = self._plot_multi(multi_scores, multi_values, multi_order, 'starting population')
+            # want maximum from first round to set base upper limit
+            for key in self.targets.keys():
+                if self.targets[key][2] < 0: # want a maximum from first round
+                    setit = 0
+                    for multi in multi_values:
+                        setit = max(multi[key], setit)
+                    self.targets[key][2] = setit
+                if self.targets[key][3] < 0: # want a maximum from first round
+                    setit = 0
+                    for multi in multi_values:
+                        setit = max(multi[key], setit)
+                    self.targets[key][3] = setit
+            # now we can find the best weighted result - lowest is best
+            best_multi = np.min(multi_scores)
+            best_mndx = multi_scores.index(best_multi)
+            multi_lowest_chrom = population[best_mndx]
+            multi_best_popn.append(multi_lowest_chrom)
+            multi_best.append(multi_values[best_mndx])
+            self.setStatus('Starting Weight: %.4f' % best_multi)
+            multi_best_weight = best_multi
+            best_multi_progress = [best_multi]
+            if not do_lcoe:
+                best_score = best_multi
+            last_multi_score = best_multi
+            lowest_multi_score = best_multi
+            mud = '='
+        # Add starting best score to progress tracker
+        best_score_progress = [best_score]
+        best_ctr = 1
+        last_score = best_score
+        lowest_score = best_score
+        lud = '='
+        # Now we'll go through the generations of genetic algorithm
+        for generation in range(1, maximum_generation):
+            lcoe_status = ''
+            multi_status = ''
+            if do_lcoe:
+                lcoe_status = ' %s $%.2f ;' % (lud, best_score)
+            if do_multi:
+                multi_status = ' %s %.4f ;' % (mud, best_multi)
+            tim = (time.time() - start_time)
+            if tim < 60:
+                tim = ' (%s%s %.1f secs)' % (lcoe_status, multi_status, tim)
+            else:
+                tim = ' (%s%s %.2f mins)' % (lcoe_status, multi_status, tim / 60.)
+            self.opt_progressbar.barProgress(generation + 1,
+                'Processing iteration ' + str(generation + 1) + tim)
+            QtWidgets.QApplication.processEvents()
+            if not self.opt_progressbar.be_open:
+                break
+        # Create an empty list for new population
+            new_population = []
+        # Using elitism approach include best individual
+            if do_lcoe:
+                new_population.append(lowest_chrom)
+            if do_multi:
+                new_population.append(multi_lowest_chrom)
+            # Create new population generating two children at a time
+            if do_lcoe:
+                if do_multi:
+                    for i in range(int(population_size/2)):
+                        parent_1 = select_individual_by_tournament(population, lcoe_scores,
+                                                                   multi_scores)
+                        parent_2 = select_individual_by_tournament(population, lcoe_scores,
+                                                                   multi_scores)
+                        child_1, child_2 = breed_by_crossover(parent_1, parent_2)
+                        new_population.append(child_1)
+                        new_population.append(child_2)
+                else:
+                    for i in range(int(population_size/2)):
+                        parent_1 = select_individual_by_tournament(population, lcoe_scores)
+                        parent_2 = select_individual_by_tournament(population, lcoe_scores)
+                        child_1, child_2 = breed_by_crossover(parent_1, parent_2)
+                        new_population.append(child_1)
+                        new_population.append(child_2)
+            else:
+                for i in range(int(population_size/2)):
+                    parent_1 = select_individual_by_tournament(population, multi_scores)
+                    parent_2 = select_individual_by_tournament(population, multi_scores)
+                    child_1, child_2 = breed_by_crossover(parent_1, parent_2)
+                    new_population.append(child_1)
+                    new_population.append(child_2)
+            # get back to original size (after elitism adds)
+            if do_lcoe:
+                new_population.pop()
+            if do_multi:
+                new_population.pop()
+            # Replace the old population with the new one
+            population = np.array(new_population)
+            if self.optimise_mutation > 0:
+                population = randomly_mutate_population(population, self.optimise_mutation)
+            # Score best solution, and add to tracker
+            lcoe_scores, multi_scores, multi_values = calculate_fitness(population)
+            if do_lcoe:
+                best_lcoe = np.min(lcoe_scores)
+                best_ndx = lcoe_scores.index(best_lcoe)
+                best_score = best_lcoe
+            # now we can find the best weighted result - lowest is best
+            if do_multi:
+                best_multi = np.min(multi_scores)
+                best_mndx = multi_scores.index(best_multi)
+                multi_lowest_chrom = population[best_mndx]
+                multi_best_popn.append(multi_lowest_chrom)
+                multi_best.append(multi_values[best_mndx])
+           #     if multi_best_weight > best_multi:
+                multi_best_weight = best_multi
+                if not do_lcoe:
+                    best_score = best_multi
+                best_multi_progress.append(best_multi)
+            best_score_progress.append(best_score)
+            if best_score < lowest_score:
+                lowest_score = best_score
+                if do_lcoe:
+                    lowest_chrom = population[best_ndx]
+                else: #(do_multi only)
+                    multi_lowest_chrom = population[best_mndx]
+            if self.optimise_stop > 0:
+                if best_score == last_score:
+                    best_ctr += 1
+                    if best_ctr >= self.optimise_stop:
+                        break
+                else:
+                    last_score = best_score
+                    best_ctr = 1
+            last_score = best_score
+            if do_lcoe:
+                if best_score == best_score_progress[-2]:
+                    lud = '='
+                elif best_score < best_score_progress[-2]:
+                    lud = '<html>&darr;</html>'
+                else:
+                    lud = '<html>&uarr;</html>'
+            if do_multi:
+                if best_multi == last_multi_score:
+                    mud = '='
+                elif best_multi < last_multi_score:
+                    mud = '<html>&darr;</html>'
+                else:
+                    mud = '<html>&uarr;</html>'
+                last_multi_score = best_multi
+        if self.debug:
+            try:
+                self.db_file.close()
+                optimiseDebug(self.db_file.name)
+                os.remove(self.db_file.name)
+            except:
+                pass
+            self.debug = False
+        self.opt_progressbar.setVisible(False)
+        self.opt_progressbar.close()
+        tim = (time.time() - start_time)
+        if tim < 60:
+            tim = '%.1f secs)' % tim
+        else:
+            tim = '%.2f mins)' % (tim / 60.)
+        msg = 'Optimise completed (%0d iterations; %s' % (generation + 1, tim)
+        if best_score > lowest_score:
+            msg += ' Try more iterations.'
+        # we'll keep two or three to save re-calculating_fitness
+        op_data = [[], [], [], [], []]
+        score_data = [None, None, None, None, None]
+        if do_lcoe:
+            op_data[0], score_data[0] = calculate_fitness([lowest_chrom])
+        if do_multi:
+            op_data[1], score_data[1] = calculate_fitness([multi_lowest_chrom])
+        self.setStatus(msg)
+        QtWidgets.QApplication.processEvents()
+        self.progressbar.setHidden(True)
+        self.progressbar.setValue(0)
+        # GA has completed required generation
+        if do_lcoe:
+            self.setStatus('Final LCOE: $%.2f' % best_score)
+            fig = 'optimise_lcoe'
+            titl = 'Optimise LCOE using Genetic Algorithm'
+            ylbl = 'Best LCOE ($/MWh)'
+        else:
+            fig = 'optimise_multi'
+            titl = 'Optimise Multi using Genetic Algorithm'
+            ylbl = 'Best Weight'
+        if do_multi:
+            self.setStatus('Final Weight: %.4f' % multi_best_weight)
+        # Plot progress
+        x = list(range(1, len(best_score_progress)+ 1))
+        matplotlib.rcParams['savefig.directory'] = self.scenarios
+        plt.figure(fig + datetime.now().strftime('_%Y-%M-%d_%H%M'))
+        lx = plt.subplot(111)
+        plt.title(titl)
+        lx.plot(x, best_score_progress)
+        lx.set_xlabel('Optimise Cycle (' + str(len(best_score_progress)) + ' iterations)')
+        lx.set_ylabel(ylbl)
+        zp = ZoomPanX()
+        f = zp.zoom_pan(lx, base_scale=1.2, annotate=True)
+        plt.show()
+        pick = None
+        pickf = None
+        if do_multi:
+            if self.optimise_multiplot:
+                pick = self._plot_multi(best_multi_progress, multi_best, multi_order, 'best of each iteration')
+                if self.more_details:
+                    pickf = self._plot_multi(multi_scores, multi_values, multi_order, 'final iteration')
+            if self.optimise_multitable:
+                pick2 = show_multitable(best_multi_progress, multi_best, multi_order, 'best of each iteration')
+                try:
+                    pick = pick + pick2
+                except:
+                    pick = pick2
+                if self.more_details:
+                    pick2 = show_multitable(multi_scores, multi_values, multi_order, 'final iteration')
+                    try:
+                        pickf = pickf + pick2
+                    except:
+                        pickf = pick2
+        op_pts = [0] * len(headers)
+        for p in [st_lcg, st_lco, st_lcc, st_max, st_bal, st_rlc, st_are]:
+            op_pts[p] = 2
+        op_pts[st_cap] = 3
+        if self.more_details:
+            if do_lcoe:
+                list(map(list, list(zip(*op_data[0]))))
+                dialog = self.displaytable.Table(op_data[0], title=sender_text, fields=headers,
+                         save_folder=self.scenarios, sortby='', decpts=op_pts)
+                dialog.exec_()
+                del dialog
+            if do_multi:
+                list(map(list, list(zip(*op_data[1]))))
+                dialog = self.displaytable.Table(op_data[1], title='Multi_' + sender_text, fields=headers,
+                         save_folder=self.scenarios, sortby='', decpts=op_pts)
+                dialog.exec_()
+                del dialog
+        # now I'll display the resulting capacities for LCOE, lowest weight, picked
+        # now get random amount of generation per technology (both RE and non-RE)
+        its = {}
+        for fac, value in opt_order.items():
+            its[fac] = []
+        chrom_hdrs = []
+        chroms = []
+        ndxes = []
+        if do_lcoe:
+            chrom_hdrs = ['Lowest LCOE']
+            chroms = [lowest_chrom]
+            ndxes = [0]
+        if do_multi:
+            chrom_hdrs.append('Lowest Weight')
+            chroms.append(multi_lowest_chrom)
+            ndxes.append(1)
+        if pickf is not None:
+            for p in range(len(pickf)):
+                if pick is None:
+                    pick = [pickf[f]]
+                else:
+                    pick.append(pickf[p][:])
+                pick[-1][0] = len(multi_best_popn)
+                multi_best_popn.append(population[pickf[p][0]])
+        if pick is not None:
+            # at present I'll calculate the best weight for the chosen picks. Could actually present all for user choice
+            if len(pick) <= 3:
+                multi_lowest_chrom = multi_best_popn[pick[0][0]]
+                op_data[2], score_data[2] = calculate_fitness([multi_lowest_chrom])
+                if self.more_details:
+                    list(map(list, list(zip(*op_data[2]))))
+                    dialog = self.displaytable.Table(op_data[2], title='Pick_' + sender_text, fields=headers,
+                             save_folder=self.scenarios, sortby='', decpts=op_pts)
+                    dialog.exec_()
+                    del dialog
+                chrom_hdrs.append('Your pick')
+                chroms.append(multi_lowest_chrom)
+                ndxes.append(2)
+                if len(pick) >= 2:
+                    multi_lowest_chrom = multi_best_popn[pick[1][0]]
+                    op_data[3], score_data[3] = calculate_fitness([multi_lowest_chrom])
+                    if self.more_details:
+                        list(map(list, list(zip(*op_data[3]))))
+                        dialog = self.displaytable.Table(op_data[3], title='Pick_' + sender_text, fields=headers,
+                                 save_folder=self.scenarios, sortby='', decpts=op_pts)
+                        dialog.exec_()
+                        del dialog
+                    chrom_hdrs.append('Your 2nd pick')
+                    chroms.append(multi_lowest_chrom)
+                    ndxes.append(3)
+                if len(pick) == 3:
+                    multi_lowest_chrom = multi_best_popn[pick[2][0]]
+                    op_data[4], score_data[4] = calculate_fitness([multi_lowest_chrom])
+                    if self.more_details:
+                        list(map(list, list(zip(*op_data[4]))))
+                        dialog = self.displaytable.Table(op_data[4], title='Pick_' + sender_text, fields=headers,
+                                 save_folder=self.scenarios, sortby='', decpts=op_pts)
+                        dialog.exec_()
+                        del dialog
+                    chrom_hdrs.append('Your 3rd pick')
+                    chroms.append(multi_lowest_chrom)
+                    ndxes.append(4)
+            else:
+                picks = []
+                for pck in pick:
+                    picks.append(multi_best_popn[pck[0]])
+                a, b, c = calculate_fitness(picks)
+                best_multi = np.min(b)
+                best_mndx = b.index(best_multi)
+                multi_lowest_chrom = picks[best_mndx]
+                op_data[2], score_data[2] = calculate_fitness([multi_lowest_chrom])
+                if self.more_details:
+                    list(map(list, list(zip(*op_data[2]))))
+                    dialog = self.displaytable.Table(op_data[2], title='Pick_' + sender_text, fields=headers,
+                             save_folder=self.scenarios, sortby='', decpts=op_pts)
+                    dialog.exec_()
+                    del dialog
+                chrom_hdrs.append('Your pick')
+                chroms.append(multi_lowest_chrom)
+                ndxes.append(2)
+        for chromosome in chroms:
+            for fac, value in opt_order.items():
+                capacity = opt_order[fac][2]
+                for c in range(value[0], value[1]):
+                    if chromosome[c]:
+                        capacity = capacity + capacities[c]
+                its[fac].append(capacity / pmss_details[fac].capacity)
+                
+        chooseDialog = QtWidgets.QDialog()
+        hbox = QtWidgets.QHBoxLayout()
+        grid = [QtWidgets.QGridLayout()]
+        label = QtWidgets.QLabel('<b>Facility</b>')
+        label.setAlignment(QtCore.Qt.AlignCenter)
+        grid[0].addWidget(label, 0, 0)
+        for h in range(len(chrom_hdrs)):
+            grid.append(QtWidgets.QGridLayout())
+            label = QtWidgets.QLabel('<b>' + chrom_hdrs[h] + '</b>')
+            label.setAlignment(QtCore.Qt.AlignCenter)
+            grid[-1].addWidget(label, 0, 0, 1, 3)
+        rw = 1
+        for key, value in its.items():
+            grid[0].addWidget(QtWidgets.QLabel(key), rw, 0)
+            if pmss_details[key].fac_type == 'S':
+                typ = ' MWh'
+            else:
+                typ = ' MW'
+            if self.show_multipliers:
+                for h in range(len(chrom_hdrs)):
+                    label = QtWidgets.QLabel('{:,.1f}'.format(value[h] * pmss_details[key].capacity))
+                    label.setAlignment(QtCore.Qt.AlignRight)
+                    grid[h + 1].addWidget(label, rw, 0)
+                    label = QtWidgets.QLabel(typ)
+                    label.setAlignment(QtCore.Qt.AlignLeft)
+                    grid[h + 1].addWidget(label, rw, 1)
+                    label = QtWidgets.QLabel('({:.2f})'.format(value[h]))
+                    label.setAlignment(QtCore.Qt.AlignRight)
+                    grid[h + 1].addWidget(label, rw, 2)
+            else:
+                for h in range(len(chrom_hdrs)):
+                    label = QtWidgets.QLabel('{:,.1f}'.format(value[h] * pmss_details[key].capacity))
+                    label.setAlignment(QtCore.Qt.AlignRight)
+                    grid[h + 1].addWidget(label, rw, 0, 1, 2)
+                    label = QtWidgets.QLabel(typ)
+                    label.setAlignment(QtCore.Qt.AlignLeft)
+                    grid[h + 1].addWidget(label, rw, 2)
+            rw += 1
+
+        max_amt = [0., 0.]
+        if do_lcoe:
+            max_amt[0] = score_data[0][0]['cost']
+            max_amt[1] = score_data[0][0]['co2']
+        if do_multi:
+            for multi in multi_best:
+                max_amt[0] = max(max_amt[0], multi['cost'])
+                max_amt[1] = max(max_amt[1], multi['co2'])
+        pwr_chr = ['', '']
+        divisor = [1., 1.]
+        pwr_chrs = ' KMBTPEZY'
+        for m in range(2):
+            for pwr in range(len(pwr_chrs) - 1, -1, -1):
+                if max_amt[m] > pow(10, pwr * 3):
+                    pwr_chr[m] = pwr_chrs[pwr]
+                    divisor[m] = 1. * pow(10, pwr * 3)
+                    break
+        self.targets['cost'][5] = self.targets['cost'][5].replace('pwr_chr', pwr_chr[0])
+        self.targets['co2'][5] = self.targets['co2'][5].replace('pwr_chr', pwr_chr[1])
+        for key in multi_order:
+            lbl = QtWidgets.QLabel('<i>' + self.targets[key][0] + '</i>')
+            lbl.setAlignment(QtCore.Qt.AlignCenter)
+            grid[0].addWidget(lbl, rw, 0)
+            for h in range(len(chrom_hdrs)):
+                if key == 'cost':
+                    amt = score_data[ndxes[h]][0][key] / divisor[0] # cost
+                elif key == 'co2':
+                    amt = score_data[ndxes[h]][0][key] / divisor[1] # co2
+                elif key[-4:] == '_pct': # percentage
+                    amt = score_data[ndxes[h]][0][key] * 100.
+                else:
+                    amt = score_data[ndxes[h]][0][key]
+                txt = '<i>' + self.targets[key][5] + '</i>'
+                try:
+                    label = QtWidgets.QLabel(txt % amt)
+                except:
+                    label = QtWidgets.QLabel('?')
+                    print('PME5:', key, txt, amt)
+                label.setAlignment(QtCore.Qt.AlignCenter)
+                grid[h + 1].addWidget(label, rw, 0, 1, 3)
+            rw += 1
+        cshow = QtWidgets.QPushButton('Quit', self)
+        grid[0].addWidget(cshow)
+        cshow.clicked.connect(chooseDialog.close)
+        for h in range(len(chrom_hdrs)):
+            button = QtWidgets.QPushButton(chrom_hdrs[h], self)
+            grid[h + 1].addWidget(button, rw, 0, 1, 3)
+            button.clicked.connect(chooseClicked) #(chrom_hdrs[h]))
+        for gri in grid:
+            frame = QtWidgets.QFrame()
+            frame.setFrameStyle(QtWidgets.QFrame.Box)
+            frame.setLineWidth(1)
+            frame.setLayout(gri)
+            hbox.addWidget(frame)
+   #     grid.addWidget(show, rw, 1)
+    #    show.clicked.connect(optDialog.close)
+        chooseDialog.setLayout(hbox)
+        chooseDialog.setWindowTitle('Choose Optimal Generator Mix')
+        chooseDialog.setWindowIcon(QtGui.QIcon('sen_icon32.ico'))
+        
+     #  this is a big of a kluge but I couldn't get it to behave
+        self.opt_choice = ''
+        chooseDialog.exec_()
+        del chooseDialog
+        try:
+            h = chrom_hdrs.index(self.opt_choice)
+        except:
+            return
+        op_data[h], score_data[h] = calculate_fitness([chroms[h]]) # make it current
+        msg = chrom_hdrs[h] + ': '
+        for key in multi_order[:3]:
+            msg += self.targets[key][0] + ': '
+            if key == 'cost':
+                amt = score_data[h][0][key] / divisor[0] # cost
+            elif key == 'co2':
+                amt = score_data[h][0][key] / divisor[1] # co2
+            elif key[-4:] == '_pct': # percentage
+                amt = score_data[h][0][key] * 100.
+            else:
+                amt = score_data[h][0][key]
+            txt = self.targets[key][5]
+            txt = txt % amt
+            msg += txt + '; '
+        self.setStatus(msg)
+        list(map(list, list(zip(*op_data[h]))))
+        op_data[h].append(' ')
+        op_data[h].append('Optimisation Parameters')
+        op_op_prm = len(op_data[h])
+        op_data[h].append(['Population size', str(self.optimise_population)])
+        op_data[h].append(['No. of iterations', str(self.optimise_generations)])
+        op_data[h].append(['Mutation probability', '%0.4f' % self.optimise_mutation])
+        op_data[h].append(['Exit if stable', str(self.optimise_stop)])
+        op_data[h].append(['Optimisation choice', self.optimise_choice])
+        op_data[h].append(['Variable', 'Weight', 'Better', 'Worse'])
+        for i in range(len(target_keys)):
+            op_data[h].append([])
+            for j in range(4):
+                if j == 0:
+                    op_data[h][-1].append(self.targets[target_keys[i]][j])
+                else:
+                    op_data[h][-1].append('{:.2f}'.format(self.targets[target_keys[i]][j]))
+        op_max_row = len(op_data[h])
+        for key in self.optimisation.keys():
+            op_data[h].append(['Max. ' + key, self.optimisation[key].capacity_max])
+        
+        dialog = Table(
+            objects=op_data[h],  # The data to display
+            fields=headers,      # The fields (columns) to display
+            parent=None          # Optional parent for the dialog
+        )
+        dialog.setWindowTitle('Chosen_' + sender_text)  # Set the title of the dialog
+        dialog.exec_()  # Execute the dialog
+        del dialog  # Clean up the dialog
+
+        if self.optimise_to_batch:
+            msgbox = QtWidgets.QMessageBox()
+            msgbox.setWindowTitle('SIREN - Add to Batch')
+            msgbox.setText("Press 'Yes' to add to Batch file")
+            msgbox.setIcon(QtWidgets.QMessageBox.Question)
+            msgbox.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
+            reply = msgbox.exec_()
+            if reply == QtWidgets.QMessageBox.Yes:
+                check_list = []
+                tot_row = 0
+                save_opt_rows = False
+                for o_r in range(len(op_data[h])):
+                    if op_data[h][o_r][0] == 'Total':
+                        break
+                    if op_data[h][o_r][0] != 'RE Contribution To Load':
+                        check_list.append(o_r)
+                if self.files[B].text() != '':
+                    newfile = self.get_filename(self.files[B].text())
+                else:
+                    curfile = self.scenarios[:-1]
+                    newfile = QtWidgets.QFileDialog.getSaveFileName(None, 'Create and save ' + self.file_labels[B] + ' file',
+                              curfile, 'Excel Files (*.xlsx)')[0]
+                    if newfile == '':
+                        return
+                if os.path.exists(newfile):
+                    wb = oxl.load_workbook(newfile)
+                elif self.batch_template == '':
+                    return
+                else:
+                    wb = oxl.load_workbook(self.batch_template)   #copy batch
+                    if newfile[: len(self.scenarios)] == self.scenarios:
+                        self.files[B].setText(newfile[len(self.scenarios):])
+                    else:
+                        if newfile.rfind('/') > 0:
+                            that_len = len(commonprefix([self.scenarios, newfile]))
+                            if that_len > 0:
+                                bits = self.scenarios[that_len:].split('/')
+                                pfx = ('..' + '/') * (len(bits) - 1)
+                                newfile = pfx + newfile[that_len + 1:]
+                        if newfile[-5:] != '.xlsx':
+                            newfile += '.xlsx'
+                        self.files[B].setText(newfile)
+                if wb.worksheets[0].max_column > 1024:
+                    self.clean_batch_sheet()
+                    ds = oxl.load_workbook(self.get_filename(self.files[B].text()))
+                batch_input_sheet = wb.worksheets[0]
+                batch_input_sheet.protection.sheet = False
+                normal = oxl.styles.Font(name='Arial')
+                bold = oxl.styles.Font(name='Arial', bold=True)
+                col = batch_input_sheet.max_column + 1
+                fst_row = -1
+                if col == 4: # possibly only chart stuff in columns 2 and 3
+                    get_out = False
+                    for col in range(3, 1, -1):
+                        for row in range(1, batch_input_sheet.max_row + 1):
+                            if batch_input_sheet.cell(row=row, column=col).value is not None:
+                                col += 1
+                                get_out = True
+                                break
+                            if batch_input_sheet.cell(row=row, column=1).value == 'Total':
+                                break
+                        if get_out:
+                            break
+                for row in range(1, batch_input_sheet.max_row + 1):
+                    if batch_input_sheet.cell(row=row, column=1).value is None:
+                        continue
+                    if batch_input_sheet.cell(row=row, column=1).value in ['Model', 'Model Label', 'Technology']:
+                        new_cell = batch_input_sheet.cell(row=row, column=col)
+                        new_cell.value = QtCore.QDateTime.toString(QtCore.QDateTime.currentDateTime(), 'MM-dd hh:mm')
+                        msg += " Added to batch as '" + new_cell.value + "' (column " + ssCol(col) + ')'
+                        continue
+                    if batch_input_sheet.cell(row=row, column=1).value == 'Capacity (MW)':
+                        fst_row = row + 1
+                        cell = batch_input_sheet.cell(row=row, column=col - 1)
+                        new_cell = batch_input_sheet.cell(row=row, column=col)
+                        new_cell.value = 'MW'
+                        if cell.has_style:
+                            new_cell.font = copy(cell.font)
+                            new_cell.border = copy(cell.border)
+                            new_cell.fill = copy(cell.fill)
+                            new_cell.number_format = copy(cell.number_format)
+                            new_cell.protection = copy(cell.protection)
+                            new_cell.alignment = copy(cell.alignment)
+                        continue
+                    if batch_input_sheet.cell(row=row, column=1).value == 'Optimisation Parameters':
+                        save_opt_rows = True
+                        break
+                    for o_r in range(len(op_data[h])):
+                        if op_data[h][o_r][0] == batch_input_sheet.cell(row=row, column=1).value:
+                            if op_data[h][o_r][0] == 'Total' and col > 2:
+                                cell = batch_input_sheet.cell(row=row, column=2)
+                            else:
+                                cell = batch_input_sheet.cell(row=row, column=col - 1)
+                            new_cell = batch_input_sheet.cell(row=row, column=col)
+                            try:
+                                new_cell.value = float(op_data[h][o_r][1])
+                            except:
+                                try:
+                                    new_cell.value = op_data[h][o_r][1]
+                                except:
+                                    pass
+                            if cell.has_style:
+                                new_cell.font = copy(cell.font)
+                                new_cell.border = copy(cell.border)
+                                new_cell.fill = copy(cell.fill)
+                                new_cell.protection = copy(cell.protection)
+                                new_cell.alignment = copy(cell.alignment)
+                                if col == 2:
+                                    new_cell.font = normal
+                                    new_cell.number_format = '#0.00'
+                                else:
+                                    new_cell.number_format = copy(cell.number_format)
+                            elif col == 2:
+                                new_cell.font = normal
+                                new_cell.number_format = '#0.00'
+                            try:
+                                i = check_list.index(o_r)
+                                del check_list[i]
+                            except:
+                                pass
+                    if batch_input_sheet.cell(row=row, column=1).value == 'Total':
+                        tot_row = row
+                if save_opt_rows: # want optimisation?
+                    for o_r in range(op_op_prm, len(op_data[h])):
+                        row += 1
+                        new_cell = batch_input_sheet.cell(row=row, column=1)
+                        new_cell.value = op_data[h][o_r][0]
+                        new_cell = batch_input_sheet.cell(row=row, column=col)
+                        try:
+                            new_cell.value = float(op_data[h][o_r][1])
+                        except:
+                            new_cell.value = op_data[h][o_r][1]
+                if len(check_list) > 0:
+                    check_list.reverse()
+                    if col > 2:
+                        cell = batch_input_sheet.cell(row=fst_row, column=2)
+                    else:
+                        cell = batch_input_sheet.cell(row=fst_row, column=col)
+                    for o_r in check_list:
+                        batch_input_sheet.insert_rows(tot_row)
+                        new_cell = batch_input_sheet.cell(row=tot_row, column=1)
+                        new_cell.value = op_data[h][o_r][0]
+                        new_cell = batch_input_sheet.cell(row=tot_row, column=col)
+                        try:
+                            new_cell.value = float(op_data[h][o_r][1])
+                        except:
+                            new_cell.value = op_data[h][o_r][1]
+                        if cell.has_style:
+                            new_cell.font = copy(cell.font)
+                            new_cell.border = copy(cell.border)
+                            new_cell.fill = copy(cell.fill)
+                            new_cell.number_format = copy(cell.number_format)
+                            new_cell.protection = copy(cell.protection)
+                            new_cell.alignment = copy(cell.alignment)
+                wb.save(self.get_filename(self.files[B].text()))
+                self.setStatus(msg)
+        
+        if self.adjust_gen:
+            self.adjustto = {}
+            for fac, value in sorted(pmss_details.items()):
+                self.adjustto[fac] = value.capacity * value.multiplier
+        return
 
     def center(self):
         frameGm = self.frameGeometry()
@@ -1358,7 +2484,7 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
             sp_pts[4] = 3 # discharge loss
             sp_pts[8] = 3 # parasitic loss
             sp_pts[11] = 3 # recharge loss
-            dialog = displaytable.Table(
+            dialog = Table(
                 objects=self.constraints,  # The data to display
                 decpts=sp_pts,  # The number of decimal points to display
                 save_folder=self.scenarios,  # Optional folder to save the table to
@@ -1393,7 +2519,7 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
                     else:
                         sp_pts.append(2)
             
-            dialog = displaytable.Table(
+            dialog = Table(
                 objects=self.generators,  # The data to display
                 decpts=sp_pts,  # The number of decimal points to display
                 save_folder=self.scenarios,  # Optional folder to save the table to
@@ -1414,7 +2540,7 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
                 except:
                     return
 
-            dialog = displaytable.Table(
+            dialog = Table(
                 objects=self.optimisation,  # The data to display
                 decpts=sp_pts,  # The number of decimal points to display
                 save_folder=self.scenarios,  # Optional folder to save the table to
@@ -1908,435 +3034,12 @@ class PowerMatch(QtWidgets.QWidget, PowerMatchBase):
                     self.setStatus(msg)
         return zp.datapoint
 
-    def _optDialog(self):
-        optDialog = QtWidgets.QDialog()
-        grid = QtWidgets.QGridLayout()
-        grid.addWidget(QtWidgets.QLabel('Adjust load'), 0, 0)
-        self.optLoad = QtWidgets.QDoubleSpinBox()
-        self.optLoad.setRange(-1, self.adjust_cap)
-        self.optLoad.setDecimals(4)
-        self.optLoad.setSingleStep(.1)
-        rw = 0
-        grid.addWidget(self.optLoad, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Multiplier for input Load'), rw, 2, 1, 3)
-        self.optLoad.setValue(pmss_details['Load'].multiplier)
-        rw += 1
-        grid.addWidget(QtWidgets.QLabel('Population size'), rw, 0)
-        optPopn = QtWidgets.QSpinBox()
-        optPopn.setRange(10, 500)
-        optPopn.setSingleStep(10)
-        optPopn.setValue(self.optimise_population)
-        optPopn.valueChanged.connect(self.changes)
-        grid.addWidget(optPopn, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Size of population'), rw, 2, 1, 3)
-        rw += 1
-        grid.addWidget(QtWidgets.QLabel('No. of iterations'), rw, 0, 1, 3)
-        optGenn = QtWidgets.QSpinBox()
-        optGenn.setRange(10, 500)
-        optGenn.setSingleStep(10)
-        optGenn.setValue(self.optimise_generations)
-        optGenn.valueChanged.connect(self.changes)
-        grid.addWidget(optGenn, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Number of iterations (generations)'), rw, 2, 1, 3)
-        rw += 1
-        grid.addWidget(QtWidgets.QLabel('Mutation probability'), rw, 0)
-        optMutn = QtWidgets.QDoubleSpinBox()
-        optMutn.setRange(0, 1)
-        optMutn.setDecimals(4)
-        optMutn.setSingleStep(0.001)
-        optMutn.setValue(self.optimise_mutation)
-        optMutn.valueChanged.connect(self.changes)
-        grid.addWidget(optMutn, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Add in mutation'), rw, 2, 1, 3)
-        rw += 1
-        grid.addWidget(QtWidgets.QLabel('Exit if stable'), rw, 0)
-        optStop = QtWidgets.QSpinBox()
-        optStop.setRange(0, 50)
-        optStop.setSingleStep(10)
-        optStop.setValue(self.optimise_stop)
-        optStop.valueChanged.connect(self.changes)
-        grid.addWidget(optStop, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Exit if LCOE/weight remains the same after this many iterations'),
-                       rw, 2, 1, 3)
-        rw += 1
-        grid.addWidget(QtWidgets.QLabel('Optimisation choice'), rw, 0)
-        optCombo = QtWidgets.QComboBox()
-        choices = ['LCOE', 'Multi', 'Both']
-        for choice in choices:
-            optCombo.addItem(choice)
-            if choice == self.optimise_choice:
-                optCombo.setCurrentIndex(optCombo.count() - 1)
-        grid.addWidget(optCombo, rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Choose type of optimisation'),
-                       rw, 2, 1, 3)
-        rw += 1
-        # for each variable name
-        grid.addWidget(QtWidgets.QLabel('Variable'), rw, 0)
-        grid.addWidget(QtWidgets.QLabel('Weight'), rw, 1)
-        grid.addWidget(QtWidgets.QLabel('Better'), rw, 2)
-        grid.addWidget(QtWidgets.QLabel('Worse'), rw, 3)
-        rw += 1
-        ndx = grid.count()
-        for key in self.targets.keys():
-            self.targets[key][4] = ndx
-            ndx += 4
-        for key, value in self.targets.items():
-            if value[2] == value[3]:
-                ud = '(=)'
-            elif value[2] < 0:
-                ud = '(<html>&uarr;</html>)'
-            elif value[3] < 0 or value[3] > value[2]:
-                ud = '(<html>&darr;</html>)'
-            else:
-                ud = '(<html>&uarr;</html>)'
-            grid.addWidget(QtWidgets.QLabel(value[0] + ': ' + ud), rw, 0)
-            weight = QtWidgets.QDoubleSpinBox()
-            weight.setRange(0, 1)
-            weight.setDecimals(2)
-            weight.setSingleStep(0.05)
-            weight.setValue(value[1])
-            grid.addWidget(weight, rw, 1)
-            if key[-4:] == '_pct':
-                minim = QtWidgets.QDoubleSpinBox()
-                minim.setRange(-.1, 1.)
-                minim.setDecimals(2)
-                minim.setSingleStep(0.1)
-                minim.setValue(value[2])
-                grid.addWidget(minim, rw, 2)
-                maxim = QtWidgets.QDoubleSpinBox()
-                maxim.setRange(-.1, 1.)
-                maxim.setDecimals(2)
-                maxim.setSingleStep(0.1)
-                maxim.setValue(value[3])
-                grid.addWidget(maxim, rw, 3)
-            else:
-                minim = QtWidgets.QLineEdit()
-                minim.setValidator(QtGui.QDoubleValidator())
-                minim.validator().setDecimals(2)
-                minim.setText(str(value[2]))
-                grid.addWidget(minim, rw, 2)
-                maxim = QtWidgets.QLineEdit()
-                maxim.setValidator(QtGui.QDoubleValidator())
-                maxim.validator().setDecimals(2)
-                maxim.setText(str(value[3]))
-                grid.addWidget(maxim, rw, 3)
-            rw += 1
-        quit = QtWidgets.QPushButton('Quit', self)
-        grid.addWidget(quit, rw, 0)
-        quit.clicked.connect(optQuitClicked)
-        show = QtWidgets.QPushButton('Proceed', self)
-        grid.addWidget(show, rw, 1)
-        show.clicked.connect(optDialog.close)
-        optDialog.setLayout(grid)
-        optDialog.setWindowTitle('Choose Optimisation Parameters')
-        optDialog.setWindowIcon(QtGui.QIcon('sen_icon32.ico'))
-        optDialog.exec_()
-        if self.optExit: # a fudge to exit
-            self.setStatus('Execution aborted.')
-            return
-        return 
-
-    def _optDialog3(self):
-        chooseDialog = QtWidgets.QDialog()
-        hbox = QtWidgets.QHBoxLayout()
-        grid = [QtWidgets.QGridLayout()]
-        label = QtWidgets.QLabel('<b>Facility</b>')
-        label.setAlignment(QtCore.Qt.AlignCenter)
-        grid[0].addWidget(label, 0, 0)
-        for h in range(len(chrom_hdrs)):
-            grid.append(QtWidgets.QGridLayout())
-            label = QtWidgets.QLabel('<b>' + chrom_hdrs[h] + '</b>')
-            label.setAlignment(QtCore.Qt.AlignCenter)
-            grid[-1].addWidget(label, 0, 0, 1, 3)
-        rw = 1
-        for key, value in its.items():
-            grid[0].addWidget(QtWidgets.QLabel(key), rw, 0)
-            if pmss_details[key].fac_type == 'S':
-                typ = ' MWh'
-            else:
-                typ = ' MW'
-            if self.show_multipliers:
-                for h in range(len(chrom_hdrs)):
-                    label = QtWidgets.QLabel('{:,.1f}'.format(value[h] * pmss_details[key].capacity))
-                    label.setAlignment(QtCore.Qt.AlignRight)
-                    grid[h + 1].addWidget(label, rw, 0)
-                    label = QtWidgets.QLabel(typ)
-                    label.setAlignment(QtCore.Qt.AlignLeft)
-                    grid[h + 1].addWidget(label, rw, 1)
-                    label = QtWidgets.QLabel('({:.2f})'.format(value[h]))
-                    label.setAlignment(QtCore.Qt.AlignRight)
-                    grid[h + 1].addWidget(label, rw, 2)
-            else:
-                for h in range(len(chrom_hdrs)):
-                    label = QtWidgets.QLabel('{:,.1f}'.format(value[h] * pmss_details[key].capacity))
-                    label.setAlignment(QtCore.Qt.AlignRight)
-                    grid[h + 1].addWidget(label, rw, 0, 1, 2)
-                    label = QtWidgets.QLabel(typ)
-                    label.setAlignment(QtCore.Qt.AlignLeft)
-                    grid[h + 1].addWidget(label, rw, 2)
-            rw += 1
-        return
-    
-    def _optChooseDialog(self):
-        for key in multi_order:
-            lbl = QtWidgets.QLabel('<i>' + self.targets[key][0] + '</i>')
-            lbl.setAlignment(QtCore.Qt.AlignCenter)
-            grid[0].addWidget(lbl, rw, 0)
-            for h in range(len(chrom_hdrs)):
-                if key == 'cost':
-                    amt = score_data[ndxes[h]][0][key] / divisor[0] # cost
-                elif key == 'co2':
-                    amt = score_data[ndxes[h]][0][key] / divisor[1] # co2
-                elif key[-4:] == '_pct': # percentage
-                    amt = score_data[ndxes[h]][0][key] * 100.
-                else:
-                    amt = score_data[ndxes[h]][0][key]
-                txt = '<i>' + self.targets[key][5] + '</i>'
-                try:
-                    label = QtWidgets.QLabel(txt % amt)
-                except:
-                    label = QtWidgets.QLabel('?')
-                    print('PME5:', key, txt, amt)
-                label.setAlignment(QtCore.Qt.AlignCenter)
-                grid[h + 1].addWidget(label, rw, 0, 1, 3)
-            rw += 1
-        cshow = QtWidgets.QPushButton('Quit', self)
-        grid[0].addWidget(cshow)
-        cshow.clicked.connect(chooseDialog.close)
-        for h in range(len(chrom_hdrs)):
-            button = QtWidgets.QPushButton(chrom_hdrs[h], self)
-            grid[h + 1].addWidget(button, rw, 0, 1, 3)
-            button.clicked.connect(chooseClicked) #(chrom_hdrs[h]))
-        for gri in grid:
-            frame = QtWidgets.QFrame()
-            frame.setFrameStyle(QtWidgets.QFrame.Box)
-            frame.setLineWidth(1)
-            frame.setLayout(gri)
-            hbox.addWidget(frame)
-   #     grid.addWidget(show, rw, 1)
-    #    show.clicked.connect(optDialog.close)
-        chooseDialog.setLayout(hbox)
-        chooseDialog.setWindowTitle('Choose Optimal Generator Mix')
-        chooseDialog.setWindowIcon(QtGui.QIcon('sen_icon32.ico'))
-        
-     #  this is a big of a kluge but I couldn't get it to behave
-        self.opt_choice = ''
-        chooseDialog.exec_()
-        del chooseDialog
-        try:
-            h = chrom_hdrs.index(self.opt_choice)
-        except:
-            return
-        op_data[h], score_data[h] = calculate_fitness([chroms[h]]) # make it current
-        msg = chrom_hdrs[h] + ': '
-        for key in multi_order[:3]:
-            msg += self.targets[key][0] + ': '
-            if key == 'cost':
-                amt = score_data[h][0][key] / divisor[0] # cost
-            elif key == 'co2':
-                amt = score_data[h][0][key] / divisor[1] # co2
-            elif key[-4:] == '_pct': # percentage
-                amt = score_data[h][0][key] * 100.
-            else:
-                amt = score_data[h][0][key]
-            txt = self.targets[key][5]
-            txt = txt % amt
-            msg += txt + '; '
-        self.setStatus(msg)
-        list(map(list, list(zip(*op_data[h]))))
-        op_data[h].append(' ')
-        op_data[h].append('Optimisation Parameters')
-        op_op_prm = len(op_data[h])
-        op_data[h].append(['Population size', str(self.optimise_population)])
-        op_data[h].append(['No. of iterations', str(self.optimise_generations)])
-        op_data[h].append(['Mutation probability', '%0.4f' % self.optimise_mutation])
-        op_data[h].append(['Exit if stable', str(self.optimise_stop)])
-        op_data[h].append(['Optimisation choice', self.optimise_choice])
-        op_data[h].append(['Variable', 'Weight', 'Better', 'Worse'])
-        for i in range(len(target_keys)):
-            op_data[h].append([])
-            for j in range(4):
-                if j == 0:
-                    op_data[h][-1].append(self.targets[target_keys[i]][j])
-                else:
-                    op_data[h][-1].append('{:.2f}'.format(self.targets[target_keys[i]][j]))
-        op_max_row = len(op_data[h])
-        for key in self.optimisation.keys():
-            op_data[h].append(['Max. ' + key, self.optimisation[key].capacity_max])
-        
-        dialog = displaytable.Table(
-            objects=op_data[h],  # The data to display
-            fields=headers,      # The fields (columns) to display
-            parent=None          # Optional parent for the dialog
-        )
-        dialog.setWindowTitle('Chosen_' + sender_text)  # Set the title of the dialog
-        dialog.exec_()  # Execute the dialog
-        del dialog  # Clean up the dialog
-
-        if self.optimise_to_batch:
-            msgbox = QtWidgets.QMessageBox()
-            msgbox.setWindowTitle('SIREN - Add to Batch')
-            msgbox.setText("Press 'Yes' to add to Batch file")
-            msgbox.setIcon(QtWidgets.QMessageBox.Question)
-            msgbox.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
-            reply = msgbox.exec_()
-            if reply == QtWidgets.QMessageBox.Yes:
-                check_list = []
-                tot_row = 0
-                save_opt_rows = False
-                for o_r in range(len(op_data[h])):
-                    if op_data[h][o_r][0] == 'Total':
-                        break
-                    if op_data[h][o_r][0] != 'RE Contribution To Load':
-                        check_list.append(o_r)
-                if self.files[B].text() != '':
-                    newfile = self.get_filename(self.files[B].text())
-                else:
-                    curfile = self.scenarios[:-1]
-                    newfile = QtWidgets.QFileDialog.getSaveFileName(None, 'Create and save ' + self.file_labels[B] + ' file',
-                              curfile, 'Excel Files (*.xlsx)')[0]
-                    if newfile == '':
-                        return
-                if os.path.exists(newfile):
-                    wb = oxl.load_workbook(newfile)
-                elif self.batch_template == '':
-                    return
-                else:
-                    wb = oxl.load_workbook(self.batch_template)   #copy batch
-                    if newfile[: len(self.scenarios)] == self.scenarios:
-                        self.files[B].setText(newfile[len(self.scenarios):])
-                    else:
-                        if newfile.rfind('/') > 0:
-                            that_len = len(commonprefix([self.scenarios, newfile]))
-                            if that_len > 0:
-                                bits = self.scenarios[that_len:].split('/')
-                                pfx = ('..' + '/') * (len(bits) - 1)
-                                newfile = pfx + newfile[that_len + 1:]
-                        if newfile[-5:] != '.xlsx':
-                            newfile += '.xlsx'
-                        self.files[B].setText(newfile)
-                if wb.worksheets[0].max_column > 1024:
-                    self.clean_batch_sheet()
-                    ds = oxl.load_workbook(self.get_filename(self.files[B].text()))
-                batch_input_sheet = wb.worksheets[0]
-                batch_input_sheet.protection.sheet = False
-                normal = oxl.styles.Font(name='Arial')
-                bold = oxl.styles.Font(name='Arial', bold=True)
-                col = batch_input_sheet.max_column + 1
-                fst_row = -1
-                if col == 4: # possibly only chart stuff in columns 2 and 3
-                    get_out = False
-                    for col in range(3, 1, -1):
-                        for row in range(1, batch_input_sheet.max_row + 1):
-                            if batch_input_sheet.cell(row=row, column=col).value is not None:
-                                col += 1
-                                get_out = True
-                                break
-                            if batch_input_sheet.cell(row=row, column=1).value == 'Total':
-                                break
-                        if get_out:
-                            break
-                for row in range(1, batch_input_sheet.max_row + 1):
-                    if batch_input_sheet.cell(row=row, column=1).value is None:
-                        continue
-                    if batch_input_sheet.cell(row=row, column=1).value in ['Model', 'Model Label', 'Technology']:
-                        new_cell = batch_input_sheet.cell(row=row, column=col)
-                        new_cell.value = QtCore.QDateTime.toString(QtCore.QDateTime.currentDateTime(), 'MM-dd hh:mm')
-                        msg += " Added to batch as '" + new_cell.value + "' (column " + ssCol(col) + ')'
-                        continue
-                    if batch_input_sheet.cell(row=row, column=1).value == 'Capacity (MW)':
-                        fst_row = row + 1
-                        cell = batch_input_sheet.cell(row=row, column=col - 1)
-                        new_cell = batch_input_sheet.cell(row=row, column=col)
-                        new_cell.value = 'MW'
-                        if cell.has_style:
-                            new_cell.font = copy(cell.font)
-                            new_cell.border = copy(cell.border)
-                            new_cell.fill = copy(cell.fill)
-                            new_cell.number_format = copy(cell.number_format)
-                            new_cell.protection = copy(cell.protection)
-                            new_cell.alignment = copy(cell.alignment)
-                        continue
-                    if batch_input_sheet.cell(row=row, column=1).value == 'Optimisation Parameters':
-                        save_opt_rows = True
-                        break
-                    for o_r in range(len(op_data[h])):
-                        if op_data[h][o_r][0] == batch_input_sheet.cell(row=row, column=1).value:
-                            if op_data[h][o_r][0] == 'Total' and col > 2:
-                                cell = batch_input_sheet.cell(row=row, column=2)
-                            else:
-                                cell = batch_input_sheet.cell(row=row, column=col - 1)
-                            new_cell = batch_input_sheet.cell(row=row, column=col)
-                            try:
-                                new_cell.value = float(op_data[h][o_r][1])
-                            except:
-                                try:
-                                    new_cell.value = op_data[h][o_r][1]
-                                except:
-                                    pass
-                            if cell.has_style:
-                                new_cell.font = copy(cell.font)
-                                new_cell.border = copy(cell.border)
-                                new_cell.fill = copy(cell.fill)
-                                new_cell.protection = copy(cell.protection)
-                                new_cell.alignment = copy(cell.alignment)
-                                if col == 2:
-                                    new_cell.font = normal
-                                    new_cell.number_format = '#0.00'
-                                else:
-                                    new_cell.number_format = copy(cell.number_format)
-                            elif col == 2:
-                                new_cell.font = normal
-                                new_cell.number_format = '#0.00'
-                            try:
-                                i = check_list.index(o_r)
-                                del check_list[i]
-                            except:
-                                pass
-                    if batch_input_sheet.cell(row=row, column=1).value == 'Total':
-                        tot_row = row
-                if save_opt_rows: # want optimisation?
-                    for o_r in range(op_op_prm, len(op_data[h])):
-                        row += 1
-                        new_cell = batch_input_sheet.cell(row=row, column=1)
-                        new_cell.value = op_data[h][o_r][0]
-                        new_cell = batch_input_sheet.cell(row=row, column=col)
-                        try:
-                            new_cell.value = float(op_data[h][o_r][1])
-                        except:
-                            new_cell.value = op_data[h][o_r][1]
-                if len(check_list) > 0:
-                    check_list.reverse()
-                    if col > 2:
-                        cell = batch_input_sheet.cell(row=fst_row, column=2)
-                    else:
-                        cell = batch_input_sheet.cell(row=fst_row, column=col)
-                    for o_r in check_list:
-                        batch_input_sheet.insert_rows(tot_row)
-                        new_cell = batch_input_sheet.cell(row=tot_row, column=1)
-                        new_cell.value = op_data[h][o_r][0]
-                        new_cell = batch_input_sheet.cell(row=tot_row, column=col)
-                        try:
-                            new_cell.value = float(op_data[h][o_r][1])
-                        except:
-                            new_cell.value = op_data[h][o_r][1]
-                        if cell.has_style:
-                            new_cell.font = copy(cell.font)
-                            new_cell.border = copy(cell.border)
-                            new_cell.fill = copy(cell.fill)
-                            new_cell.number_format = copy(cell.number_format)
-                            new_cell.protection = copy(cell.protection)
-                            new_cell.alignment = copy(cell.alignment)
-                wb.save(self.get_filename(self.files[B].text()))
-                self.setStatus(msg)
-        return
-
     def _optQuitClicked(self):
         self.optExit = True
-        optDialog.close()
+        self.optDialog.close()
 
     def _display_table(self, objects, fields=None, title=None, save_folder='', sortby=None, decpts=None):
-        dialog = displaytable.Table(objects, title=title, fields=fields,
+        dialog = Table(objects, title=title, fields=fields,
                  save_folder=save_folder, sortby=sortby, decpts=decpts)
         dialog.exec_()
         del dialog
